Reorganize backend tests
