+++ tests/backend_validation/test_validate_backends.py
@@ -1,18 +1,5 @@
 from databuilder.backends.base import BaseBackend
 
-from ..lib.mock_backend import MockBackend
-
-
-def test_backend_tables():
-    """Test that a backend registers its table names"""
-
-    assert set(MockBackend.tables) == {
-        "patients",
-        "practice_registrations",
-        "clinical_events",
-        "positive_tests",
-    }
-
 
 def test_validate_all_backends():
     """
